+++ CHANGELOG.md
@@ -9,7 +9,7 @@ SPDX-License-Identifier: CC0-1.0
 
 [Unreleased]: https://github.com/trussed-dev/trussed-rsa-backend/compare/v0.3.0...HEAD
 
--
+- Move `RsaImportFormat` and `RsaPublicParts` to the `trussed-rsa-types` crate.
 
 ## [v0.3.0][] (2025-07-31)